requirements.txt
flask>=3.0
trafilatura>=1.6
google-genai>=1.0
python-dotenv>=1.0
gunicorn>=21.0
